+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - added default role for new users
 - added pagination on some views
 - security: restrict permissions check to context org
+- added cache management
 - upgrade to hibernate 4.3
 - lots of markup and style changes 
 - other bugfixes